m6800,h8: shorthand variable types

+++ src/devices/cpu/h8/h8.h
@@ -46,7 +46,7 @@ public:
 	template<int Sci> void sci_clk_w(int state) { m_sci[Sci]->do_clk_w(state); }
 
 	void nvram_set_battery(int state) { m_nvram_battery = bool(state); } // default is 1 (nvram_enable_backup needs to be true)
-	void nvram_set_default_value(uint8_t val) { m_nvram_defval = val; } // default is 0
+	void nvram_set_default_value(u16 val) { m_nvram_defval = val; } // default is 0
 	auto standby_cb() { return m_standby_cb.bind(); } // notifier (not an output pin)
 	int standby() { return suspended(SUSPEND_REASON_CLOCK) ? 1 : 0; }
 
@@ -61,7 +61,7 @@ public:
 
 	u16 do_read_adc(int port) { return m_read_adc[port](); }
 	u8 do_read_port(int port) { return m_read_port[port](); }
-	void do_write_port(int port, uint8_t data, uint8_t ddr) { m_write_port[port](0, data, ddr); }
+	void do_write_port(int port, u8 data, u8 ddr) { m_write_port[port](0, data, ddr); }
 	void do_sci_tx(int sci, int state) { m_sci_tx[sci](state); }
 	void do_sci_clk(int sci, int state) { m_sci_clk[sci](state); }
 
@@ -106,7 +106,7 @@ protected:
 		EXR_I  = 0x07
 	};
 
-	h8_device(const machine_config &mconfig, device_type type, const char *tag, device_t *owner, uint32_t clock, address_map_constructor map_delegate);
+	h8_device(const machine_config &mconfig, device_type type, const char *tag, device_t *owner, u32 clock, address_map_constructor map_delegate);
 
 	// device_t implementation
 	virtual void device_config_complete() override;
@@ -115,9 +115,9 @@ protected:
 
 	// device_execute_interface implementation
 	virtual bool cpu_is_interruptible() const override { return true; }
-	virtual uint32_t execute_min_cycles() const noexcept override { return 2; }
-	virtual uint32_t execute_max_cycles() const noexcept override { return 12; }
-	virtual uint32_t execute_input_lines() const noexcept override { return 0; }
+	virtual u32 execute_min_cycles() const noexcept override { return 2; }
+	virtual u32 execute_max_cycles() const noexcept override { return 12; }
+	virtual u32 execute_input_lines() const noexcept override { return 0; }
 	virtual bool execute_input_edge_triggered(int inputnum) const noexcept override { return inputnum == INPUT_LINE_NMI; }
 	virtual void execute_run() override;
 
@@ -140,7 +140,7 @@ protected:
 	address_space_config m_program_config;
 	memory_access<32, 1, 0, ENDIANNESS_BIG>::cache m_cache;
 	memory_access<32, 1, 0, ENDIANNESS_BIG>::specific m_program;
-	optional_shared_ptr<uint16_t> m_internal_ram; // for nvram
+	optional_shared_ptr<u16> m_internal_ram; // for nvram
 	devcb_read16::array<8> m_read_adc;
 	devcb_read8::array<PORT_COUNT> m_read_port;
 	devcb_write8::array<PORT_COUNT> m_write_port;
@@ -154,18 +154,18 @@ protected:
 	int m_current_dma;
 	h8_dtc_state *m_current_dtc;
 
-	uint32_t  m_PPC;           // previous program counter
-	uint32_t  m_NPC;           // next start-of-instruction program counter
-	uint32_t  m_PC;            // program counter
-	uint16_t  m_PIR;           // Prefetched word
-	uint16_t  m_IR[5];         // Fetched instruction
-	uint16_t  m_R[16];         // Rn (0-7), En (8-15, h8-300h+)
-	uint8_t   m_EXR;           // Interrupt/trace register (h8s/2000+)
-	uint8_t   m_CCR;           // Condition-code register
-	int64_t   m_MAC;           // Multiply accumulator (h8s/2600+)
-	uint8_t   m_MACF;          // MAC flags (h8s/2600+)
-	uint32_t  m_TMP1, m_TMP2;
-	uint32_t  m_TMPR;          // For debugger ER register import
+	u32  m_PPC;             // previous program counter
+	u32  m_NPC;             // next start-of-instruction program counter
+	u32  m_PC;              // program counter
+	u16  m_PIR;             // Prefetched word
+	u16  m_IR[5];           // Fetched instruction
+	u16  m_R[16];           // Rn (0-7), En (8-15, h8-300h+)
+	u8   m_EXR;             // Interrupt/trace register (h8s/2000+)
+	u8   m_CCR;             // Condition-code register
+	s64   m_MAC;            // Multiply accumulator (h8s/2600+)
+	u8   m_MACF;            // MAC flags (h8s/2600+)
+	u32  m_TMP1, m_TMP2;
+	u32  m_TMPR;            // For debugger ER register import
 
 	bool m_has_exr, m_has_mac, m_has_trace, m_supports_advanced, m_mode_advanced, m_mode_a20, m_mac_saturating;
 	bool m_has_hc; // GT913's CCR bit 5 is I, not H
@@ -176,28 +176,28 @@ protected:
 	int m_irq_level, m_taken_irq_level;
 	bool m_irq_required, m_irq_nmi;
 	bool m_standby_pending;
-	uint16_t m_nvram_defval;
+	u16 m_nvram_defval;
 	bool m_nvram_battery;
 
 	virtual void do_exec_full();
 	virtual void do_exec_partial();
-	static void add_event(uint64_t &event_time, uint64_t new_event);
+	static void add_event(u64 &event_time, u64 new_event);
 	virtual bool exr_in_stack() const;
 	virtual void update_irq_filter() = 0;
 	virtual void interrupt_taken() = 0;
-	virtual void internal_update(uint64_t current_time) = 0;
-	void recompute_bcount(uint64_t event_time);
+	virtual void internal_update(u64 current_time) = 0;
+	void recompute_bcount(u64 event_time);
 	virtual int trace_setup();
 	virtual int trapa_setup();
 	virtual void irq_setup() = 0;
 
-	virtual uint16_t read16i(uint32_t adr);
-	virtual uint8_t read8(uint32_t adr);
-	virtual void write8(uint32_t adr, uint8_t data);
-	virtual uint16_t read16(uint32_t adr);
-	virtual void write16(uint32_t adr, uint16_t data);
+	virtual u16 read16i(u32 adr);
+	virtual u8 read8(u32 adr);
+	virtual void write8(u32 adr, u8 data);
+	virtual u16 read16(u32 adr);
+	virtual void write16(u32 adr, u16 data);
 	virtual void internal(int cycles);
-	void prefetch_switch(uint32_t pc, uint16_t ir) { m_NPC = pc & 0xffffff; m_PC = pc+2; m_PIR = ir; }
+	void prefetch_switch(u32 pc, u16 ir) { m_NPC = pc & 0xffffff; m_PC = pc+2; m_PIR = ir; }
 	void prefetch_done();
 	void prefetch_done_noirq();
 	void prefetch_done_notrace();
@@ -207,104 +207,104 @@ protected:
 	u8 port_default_r(int port);
 	void port_default_w(int port, u8 data);
 
-	uint8_t do_addx8(uint8_t a, uint8_t b);
-	uint8_t do_subx8(uint8_t a, uint8_t b);
+	u8 do_addx8(u8 a, u8 b);
+	u8 do_subx8(u8 a, u8 b);
 
-	uint8_t do_inc8(uint8_t a, uint8_t b);
-	uint16_t do_inc16(uint16_t a, uint16_t b);
-	uint32_t do_inc32(uint32_t a, uint32_t b);
+	u8 do_inc8(u8 a, u8 b);
+	u16 do_inc16(u16 a, u16 b);
+	u32 do_inc32(u32 a, u32 b);
 
-	uint8_t do_add8(uint8_t a, uint8_t b);
-	uint16_t do_add16(uint16_t a, uint16_t b);
-	uint32_t do_add32(uint32_t a, uint32_t b);
+	u8 do_add8(u8 a, u8 b);
+	u16 do_add16(u16 a, u16 b);
+	u32 do_add32(u32 a, u32 b);
 
-	uint8_t do_dec8(uint8_t a, uint8_t b);
-	uint16_t do_dec16(uint16_t a, uint16_t b);
-	uint32_t do_dec32(uint32_t a, uint32_t b);
+	u8 do_dec8(u8 a, u8 b);
+	u16 do_dec16(u16 a, u16 b);
+	u32 do_dec32(u32 a, u32 b);
 
-	uint8_t do_sub8(uint8_t a, uint8_t b);
-	uint16_t do_sub16(uint16_t a, uint16_t b);
-	uint32_t do_sub32(uint32_t a, uint32_t b);
+	u8 do_sub8(u8 a, u8 b);
+	u16 do_sub16(u16 a, u16 b);
+	u32 do_sub32(u32 a, u32 b);
 
-	uint8_t do_shal8(uint8_t v);
-	uint16_t do_shal16(uint16_t v);
-	uint32_t do_shal32(uint32_t v);
+	u8 do_shal8(u8 v);
+	u16 do_shal16(u16 v);
+	u32 do_shal32(u32 v);
 
-	uint8_t do_shar8(uint8_t v);
-	uint16_t do_shar16(uint16_t v);
-	uint32_t do_shar32(uint32_t v);
+	u8 do_shar8(u8 v);
+	u16 do_shar16(u16 v);
+	u32 do_shar32(u32 v);
 
-	uint8_t do_shll8(uint8_t v);
-	uint16_t do_shll16(uint16_t v);
-	uint32_t do_shll32(uint32_t v);
+	u8 do_shll8(u8 v);
+	u16 do_shll16(u16 v);
+	u32 do_shll32(u32 v);
 
-	uint8_t do_shlr8(uint8_t v);
-	uint16_t do_shlr16(uint16_t v);
-	uint32_t do_shlr32(uint32_t v);
+	u8 do_shlr8(u8 v);
+	u16 do_shlr16(u16 v);
+	u32 do_shlr32(u32 v);
 
-	uint8_t do_rotl8(uint8_t v);
-	uint16_t do_rotl16(uint16_t v);
-	uint32_t do_rotl32(uint32_t v);
+	u8 do_rotl8(u8 v);
+	u16 do_rotl16(u16 v);
+	u32 do_rotl32(u32 v);
 
-	uint8_t do_rotr8(uint8_t v);
-	uint16_t do_rotr16(uint16_t v);
-	uint32_t do_rotr32(uint32_t v);
+	u8 do_rotr8(u8 v);
+	u16 do_rotr16(u16 v);
+	u32 do_rotr32(u32 v);
 
-	uint8_t do_rotxl8(uint8_t v);
-	uint16_t do_rotxl16(uint16_t v);
-	uint32_t do_rotxl32(uint32_t v);
+	u8 do_rotxl8(u8 v);
+	u16 do_rotxl16(u16 v);
+	u32 do_rotxl32(u32 v);
 
-	uint8_t do_rotxr8(uint8_t v);
-	uint16_t do_rotxr16(uint16_t v);
-	uint32_t do_rotxr32(uint32_t v);
+	u8 do_rotxr8(u8 v);
+	u16 do_rotxr16(u16 v);
+	u32 do_rotxr32(u32 v);
 
-	uint8_t do_shal2_8(uint8_t v);
-	uint16_t do_shal2_16(uint16_t v);
-	uint32_t do_shal2_32(uint32_t v);
+	u8 do_shal2_8(u8 v);
+	u16 do_shal2_16(u16 v);
+	u32 do_shal2_32(u32 v);
 
-	uint8_t do_shar2_8(uint8_t v);
-	uint16_t do_shar2_16(uint16_t v);
-	uint32_t do_shar2_32(uint32_t v);
+	u8 do_shar2_8(u8 v);
+	u16 do_shar2_16(u16 v);
+	u32 do_shar2_32(u32 v);
 
-	uint8_t do_shll2_8(uint8_t v);
-	uint16_t do_shll2_16(uint16_t v);
-	uint32_t do_shll2_32(uint32_t v);
+	u8 do_shll2_8(u8 v);
+	u16 do_shll2_16(u16 v);
+	u32 do_shll2_32(u32 v);
 
-	uint8_t do_shlr2_8(uint8_t v);
-	uint16_t do_shlr2_16(uint16_t v);
-	uint32_t do_shlr2_32(uint32_t v);
+	u8 do_shlr2_8(u8 v);
+	u16 do_shlr2_16(u16 v);
+	u32 do_shlr2_32(u32 v);
 
-	uint8_t do_rotl2_8(uint8_t v);
-	uint16_t do_rotl2_16(uint16_t v);
-	uint32_t do_rotl2_32(uint32_t v);
+	u8 do_rotl2_8(u8 v);
+	u16 do_rotl2_16(u16 v);
+	u32 do_rotl2_32(u32 v);
 
-	uint8_t do_rotr2_8(uint8_t v);
-	uint16_t do_rotr2_16(uint16_t v);
-	uint32_t do_rotr2_32(uint32_t v);
+	u8 do_rotr2_8(u8 v);
+	u16 do_rotr2_16(u16 v);
+	u32 do_rotr2_32(u32 v);
 
-	uint8_t do_rotxl2_8(uint8_t v);
-	uint16_t do_rotxl2_16(uint16_t v);
-	uint32_t do_rotxl2_32(uint32_t v);
+	u8 do_rotxl2_8(u8 v);
+	u16 do_rotxl2_16(u16 v);
+	u32 do_rotxl2_32(u32 v);
 
-	uint8_t do_rotxr2_8(uint8_t v);
-	uint16_t do_rotxr2_16(uint16_t v);
-	uint32_t do_rotxr2_32(uint32_t v);
+	u8 do_rotxr2_8(u8 v);
+	u16 do_rotxr2_16(u16 v);
+	u32 do_rotxr2_32(u32 v);
 
-	void set_nzv8(uint8_t v);
-	void set_nzv16(uint16_t v);
-	void set_nzv32(uint32_t v);
+	void set_nzv8(u8 v);
+	void set_nzv16(u16 v);
+	void set_nzv32(u32 v);
 
-	void set_nz16(uint16_t v);
-	void set_nz32(uint32_t v);
+	void set_nz16(u16 v);
+	void set_nz32(u32 v);
 
-	inline void r8_w(int reg, uint8_t val) {
+	inline void r8_w(int reg, u8 val) {
 		if(reg & 8)
 			m_R[reg & 7] = (m_R[reg & 7] & 0xff00) | val;
 		else
 			m_R[reg & 7] = (m_R[reg & 7] & 0xff) | (val << 8);
 	}
 
-	inline uint8_t r8_r(int reg) {
+	inline u8 r8_r(int reg) {
 		if(reg & 8)
 			return m_R[reg & 7];
 		else
@@ -319,8 +319,8 @@ protected:
 	// and the h8-300h is r32 of course, we have to be careful to mask
 	// in h8.lst there if the top bit is 1.
 
-	inline void r16_w(int reg, uint16_t val) { m_R[reg & 0xf] = val; }
-	inline uint16_t r16_r(int reg) { return m_R[reg & 0xf]; }
+	inline void r16_w(int reg, u16 val) { m_R[reg & 0xf] = val; }
+	inline u16 r16_r(int reg) { return m_R[reg & 0xf]; }
 
 #define O(o) void o ## _full(); void o ## _partial()
 	O(add_b_imm8_r8u); O(add_b_r8h_r8l); O(add_w_r16h_r16l);
@@ -432,7 +432,7 @@ protected:
 };
 
 enum {
-	H8_PC  = 1,
+	H8_PC = 1,
 	H8_R0,
 	H8_R1,
 	H8_R2,
